+++ src/main/java/entity/UserInfo.java
@@ -1,12 +1,11 @@
 package entity;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Date;
 
 public class UserInfo implements Serializable {
 
-    private BigDecimal USER_ID;
+    private Integer USER_ID;
     private String USER_NAME;
     private String USER_PASS;
     private String PHONE_NO;
@@ -24,7 +23,7 @@ public class UserInfo implements Serializable {
     private String QQ;
     public UserInfo(){}
     //更新个人信息
-    public UserInfo(String USER_NAME,String PHONE_NO,String TRUE_NAME,String GENDER,String EMAIL,String ADDRESS,String BLOOD_TYPE,Date BIRTHDAY,String QQ,String INTRO,BigDecimal USER_ID){
+    public UserInfo(String USER_NAME,String PHONE_NO,String TRUE_NAME,String GENDER,String EMAIL,String ADDRESS,String BLOOD_TYPE,Date BIRTHDAY,String QQ,String INTRO,Integer USER_ID){
         this.USER_ID=USER_ID;
         this.USER_NAME=USER_NAME;
         this.PHONE_NO=PHONE_NO;
@@ -61,11 +60,11 @@ public class UserInfo implements Serializable {
         this.QQ = QQ;
     }
 
-    public BigDecimal getUSER_ID() {
+    public Integer getUSER_ID() {
         return USER_ID;
     }
 
-    public void setUSER_ID(BigDecimal USER_ID) {
+    public void setUSER_ID(Integer USER_ID) {
         this.USER_ID = USER_ID;
     }
